requirements.txt
# BT Bot Dependencies

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
supabase>=2.0.0

# AI Services
openai>=1.3.0

# Email Parsing
beautifulsoup4>=4.12.0
html2text>=2024.2.26
lxml>=4.9.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Development
pytest>=7.4.0
httpx>=0.25.0
